2.1.3

+++ bedrock/bedrock-v313/src/main/java/com/nukkitx/protocol/bedrock/v313/serializer/ResourcePackClientResponseSerializer_v313.java
@@ -21,7 +21,7 @@ public class ResourcePackClientResponseSerializer_v313 implements PacketSerializ
         buffer.writeByte(packet.getStatus().ordinal());
 
         List<String> packIds = packet.getPackIds();
-        buffer.writeIntLE(packIds.size());
+        buffer.writeShortLE(packIds.size());
 
         for (String packId : packIds) {
             BedrockUtils.writeString(buffer, packId);
